queuebreak_public_repo/requirements.txt
pyyaml>=6.0
numpy>=1.26
pandas>=2.2
pyarrow>=16.1
matplotlib>=3.8
tqdm>=4.66
rich>=13.7
psutil>=5.9
nvidia-ml-py>=12.555.43
transformers>=4.45.0
accelerate>=0.33.0
datasets>=2.20.0
safetensors>=0.4.3
huggingface-hub>=0.24.0
